Initial commit
This is a basic camera app, intended as an exploration of dependency injection (using Dagger2)
and its effect on unit testing. It relies heavily on interfaces and concrete implementations, partly
as an alternative to mocking in instrumentation tests, but also to allow swappable components
in the running app itself (e.g. CameraSystem variations which manage the two Android camera APIs).

Currently the app creates a preview surface and a controls UI, hooks them up to a camera manager object
through a presenter, and allows the user to take a photo with the shutter button. This is processed into a
bitmap and displayed as a small preview on the display. The main activity handles lifecycle events by delegating
to the presenter, which manages starting/stopping the camera and releasing its resources.

The next steps are:
  - implementing image saving (writing to storage, registering with the media provider, EXIF etc)
  - writing unit and instrumentation tests