Add tests, implement basic photo saving, some callback tweaks
Added a PhotoStorage interface/implementation which handles saving bitmaps
and adding them to the media store.

Implemented unit and instrumentation tests for most concrete classes. Some of the
functionality relies on calls to the Android system which makes them hard to test, e.g.
checking that the Media Scanner has been invoked for newly saved images, or checking
granted permissions in various circumstances. This might need a new testing library, or
some other approach (like injecting wrappers for the framework's static calls)

Reworked listener handling in CameraPresenter - now the CameraSystem only calls the
CameraPresenter on events like taking a photo, and the presenter performs any tasks it
needs to before passing on the result to its own listener. Having the Activity put a result
listener directly in the CameraSystem was too brittle and forced all the pre-callback logic
to be done in the CameraSystem itself.